separation de la logique et de la vue pour l'index du formulaire d'inscription au webinaire
